refactor: rename ai-spacing content-script/prompt-classifier by runtime
ai-spacing/content-script.ts becomes in-content-script.ts and
ai-spacing/prompt-classifier.ts becomes in-service-worker.ts, naming
the two runtime-context halves by where they run instead of
duplicating the root entry basenames (a second service-worker.ts
read as a second service worker). Updates the two imports, the
code-comment mentions, and the docs mention. Structural only, no
behavior change.
