+++ browser-extensions/chrome/src/ai-spacing/in-service-worker.ts
