+++ browser-extensions/chrome/src/options.ts
@@ -6,7 +6,7 @@ import { playSound } from './ui/sounds';
 // Chrome logs "No output language was specified in a LanguageModel API request" for any call that declares none, and collects it into the extension's Errors page, which is a user-facing surface. The
 // warning is logged for extension pages only, never for the service worker (measured 2026-09-02), so declaring a language on this page's two calls is what clears it. Nothing here is misattested:
 // neither call ever produces model output, one probes availability and the other exists only to start the browser-wide download. The classifier, which is the session that actually prompts, stays
-// undeclared on purpose -- see ai-spacing/prompt-classifier.ts. `en` is a member of the supported set (en/ja/es/de/fr); no zh variant is, and declaring one makes availability() report unavailable.
+// undeclared on purpose -- see ai-spacing/in-service-worker.ts. `en` is a member of the supported set (en/ja/es/de/fr); no zh variant is, and declaring one makes availability() report unavailable.
 const PAGE_MODEL_LANGUAGES: LanguageModelExpected[] = [{ type: 'text', languages: ['en'] }];
 
 // Builds a Partial<Settings> for the list picked by filter mode without a computed-key cast
