+++ docs/ai-spacing.md
@@ -65,7 +65,7 @@ Labels zip against the candidates by index. Every candidate whose label passes i
 
 ### Classifier and service worker (`browser-extensions/chrome`)
 
-Each ambiguous shape's `PromptSpec` carries its shipping prompt bytes in an extension-local constant file, tagged with the `version` that earned them: the system prompt, the markless question template (an untouched sentence, with the question pointing at the symbol by quoting the character before it), the display tokens, and the canonical menu order. `prompt-classifier.ts` holds the registry, keyed by `kind`, and keeps one base session per kind in a `Map<kind, Promise<LanguageModel>>`, since each shape's system prompt rides in its own session's `initialPrompts`. A session is created at `temperature 0` and `topK 1`, cloned per candidate through a sequential loop, and recreated when the MV3 worker wakes. What is cached is the promise rather than the session, so batches that arrive while one `create()` is in flight share it, and a rejected `create()` drops that kind's entry, because the model can arrive later through the options-page download. A message whose `kind` has no prompt spec registered answers `{ok: false}`, the same clean no the page gets for an absent model.
+Each ambiguous shape's `PromptSpec` carries its shipping prompt bytes in an extension-local constant file, tagged with the `version` that earned them: the system prompt, the markless question template (an untouched sentence, with the question pointing at the symbol by quoting the character before it), the display tokens, and the canonical menu order. `ai-spacing/in-service-worker.ts` holds the registry, keyed by `kind`, and keeps one base session per kind in a `Map<kind, Promise<LanguageModel>>`, since each shape's system prompt rides in its own session's `initialPrompts`. A session is created at `temperature 0` and `topK 1`, cloned per candidate through a sequential loop, and recreated when the MV3 worker wakes. What is cached is the promise rather than the session, so batches that arrive while one `create()` is in flight share it, and a rejected `create()` drops that kind's entry, because the model can arrive later through the options-page download. A message whose `kind` has no prompt spec registered answers `{ok: false}`, the same clean no the page gets for an absent model.
 
 Any `create()` failure yields `{ok: false}` with no retry arm, which is a correctness decision rather than tidiness: the sampling knobs are extension-context Chrome 151+ while the manifest floor is 99, so a retry that dropped them would silently run model-default sampling on older Chrome — the exact drift that flipped control cases in the plain-page measurements. A context that cannot pin sampling yields no session at all.
